Fix "buildhistorylinks" management command.
Previously, it was relying on NoArgsCommand. This was removed in Django 1.10
because `BaseCommand` does the same thing.

While I'm there, add a test for it so it can't get quietly broken again.

Driveby: `BaseCommandstdout.write` automatically adds line endings (since at
least Django 1.8), so there's no need to do that by hand.