Fix HistoryLinkFallbackMiddleware on Django 4.2.
Previously, the middleware was (mis?)using request.META to store the flag
to determine if it was active or not. This interfered with Django's header
handling, e.g. if you did this:

```
request.headers.get('Some-Header')
```

you would get this from deep in the bowels of Django when it saw the tuple
which was being used as the key:

```
AttributeError: 'tuple' object has no attribute 'startswith'
```

Instead, let's be less fancy. Set an attribute on the request when we are
active, and unset it when we exit. That it no longer adds `self` into the
unique key presumably means it'll break some mad use case like adding
multiple copies of the middleware. Oh well!

While we're there, add 4.2 to the test matrix, and remove the 4.0 series
as they are no longer supported.